DEBUGGING: model switching is achievedgit log! Still needs some testing and cleaning...

+++ ocs2_ros_interfaces/include/ocs2_ros_interfaces/mrt/MRT_ROS_Gazebo_Loop.h
@@ -308,7 +308,7 @@ class MRT_ROS_Gazebo_Loop
 
     bool dataCollectionFlag_ = true;
 
-    double initPolicyCtrMax_ = 2000;
+    int initPolicyCtrMax_ = 2000;
 
     double err_threshold_pos_;
     double err_threshold_ori_yaw_;
